requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
Django==3.0.8
django-common-helpers==0.9.2
django-compat==1.0.15
django-cron==0.5.1
django-transaction-hooks==0.3
lxml==4.5.1
mysql-connector==2.2.9
mysql-connector-python==8.0.20
mysqlclient==2.0.1
parse==1.15.0
parso==0.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
requests==2.25.1
requests-html==0.10.0
urllib3==1.25.9
websockets==8.1
django-meta==1.7.0
xmltodict3==0.0.4